requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.34.1
anyio==4.4.0
asttokens==2.4.1
attrs==24.2.0
boto3==1.34.162
botocore==1.34.162
certifi==2024.7.4
charset-normalizer==3.3.2
comm==0.2.2
dataclasses-json==0.6.7
decorator==5.1.1
distro==1.9.0
executing==2.0.1
faiss-cpu==1.8.0.post1
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.6
idna==3.7
ipython==8.26.0
ipywidgets==8.1.5
jedi==0.19.1
jiter==0.5.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyterlab_widgets==3.0.13
langchain==0.2.14
langchain-aws==0.1.17
langchain-community==0.2.12
langchain-core==0.2.34
langchain-text-splitters==0.2.2
langsmith==0.1.103
marshmallow==3.22.0
matplotlib-inline==0.1.7
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.7
packaging==24.1
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
s3transfer==0.10.2
safetensors==0.4.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
stack-data==0.6.3
tenacity==8.5.0
tokenizers==0.19.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==1.26.19
wcwidth==0.2.13
widgetsnbextension==4.0.13
yarl==1.9.4
